Update Maize to match Chia 1.6.0

+++ maize/ssl/create_ssl.py
@@ -24,7 +24,7 @@ def get_mozilla_ca_crt() -> str:
     return str(mozilla_path)
 
 
-def write_ssl_cert_and_key(cert_path: Path, cert_data: bytes, key_path: Path, key_data: bytes):
+def write_ssl_cert_and_key(cert_path: Path, cert_data: bytes, key_path: Path, key_data: bytes, overwrite: bool = True):
     flags = os.O_CREAT | os.O_EXCL | os.O_WRONLY
 
     for path, data, mode in [
@@ -32,6 +32,9 @@ def write_ssl_cert_and_key(cert_path: Path, cert_data: bytes, key_path: Path, ke
         (key_path, key_data, DEFAULT_PERMISSIONS_KEY_FILE),
     ]:
         if path.exists():
+            if not overwrite:
+                continue
+
             path.unlink()
 
         with open(os.open(str(path), flags, mode), "wb") as f:
@@ -68,7 +71,7 @@ def generate_ca_signed_cert(ca_crt: bytes, ca_key: bytes, cert_out: Path, key_ou
         .not_valid_before(datetime.datetime.today() - one_day)
         .not_valid_after(datetime.datetime(2100, 8, 2))
         .add_extension(
-            x509.SubjectAlternativeName([x509.DNSName("maize.farm")]),
+            x509.SubjectAlternativeName([x509.DNSName("maize.net")]),
             critical=False,
         )
         .sign(root_key, hashes.SHA256(), default_backend())
